Now InlineRenderer and WorkerRenderer extend of MMLBuilder, calling to .toXML method when use_worker is true, it creates a new process and sends MMLBuilder's context to child process to regenerate MMLBuilder and render the XML, then child process sends the XML back to the main process. InlineRenderer is just a dummy wrapper of MMLBuilder
