+++ lib/grainstore/mml_builder.js
@@ -1,7 +1,9 @@
 var _      = require('underscore');
 var millstone = require('millstone');
 var StyleTrans = require('./style_trans');
-var InlineRenderer = require('./renderer/inline-renderer');
+var carto = require('carto');
+var debug = require('debug')('grainstore:renderer');
+var debugXml = require('debug')('grainstore:renderer:xml');
 
 // configure grainstore from optional args passed + defaults
 var grainstore_defaults = {
@@ -80,7 +82,6 @@ var supportedDatasourceTypes = {
 //     }
 //
 function MMLBuilder(params, options) {
-    this.renderer = options.renderer || new InlineRenderer();
     this.params = params || {};
     // core variables
     var requiredParams = ['dbname', 'sql', 'style'];
@@ -209,7 +210,21 @@ MMLBuilder.prototype.render = function(style_in, version, callback){
         var carto_env = _.defaults({}, self.options.carto_env);
         var carto_options = { mapnik_version: self.target_mapnik_version };
 
-        self.renderer.getXML(mml, carto_options, carto_env, callback);
+        // carto.Renderer may throw during parse time (before nextTick is called)
+        // See https://github.com/mapbox/carto/pull/187
+        try {
+            var r = new carto.Renderer(carto_env, carto_options);
+            debug('Renderer.render start');
+
+            var xml = r.render(mml);
+            debugXml('output', xml);
+
+            process.nextTick(function () {
+                return callback(null, xml);
+            });
+        } catch (err) {
+            return callback(err, null);
+        }
     });
 };
