+++ lib/grainstore/mml_store.js
@@ -1,4 +1,3 @@
-var MMLBuilder = require('./mml_builder');
 var fs = require('fs');
 var _ = require('underscore');
 var debug = require('debug')('grainstore:mmlstore');
@@ -19,11 +18,16 @@ function MMLStore(options) {
 
   options = options || {};
   options.cachedir = options.cachedir || '/tmp/millstone';
-  options.renderer = options.use_workers ? new WorkerRenderer() : new InlineRenderer();
 
   // @param callback(err, payload) called on initialization
   me.mml_builder = function(params, overrideOptions) {
-    return new MMLBuilder(params, _.extend({}, options, overrideOptions));
+    var opts = _.extend({}, options, overrideOptions);
+
+    if (options.use_workers) {
+        return new WorkerRenderer(params, opts);
+    }
+
+    return new InlineRenderer(params, opts);
   };
 
   /// API: Purge cache of localized resources for this store
