+++ lib/grainstore/renderer/carto-renderer-child.js
@@ -1,15 +1,22 @@
-var carto  = require('carto');
+var _ = require('underscore');
 var debug = require('debug')('grainstore:carto-renderer-child');
-
 var InlineRenderer = require('./inline-renderer');
-var renderer = new InlineRenderer();
 
 process.on('message', function(input) {
     debug('Child got Mapnik MML');
-    renderer.getXML(input.mml, input.options, input.env, function(err, xml) {
-        if (err) {
-            return process.send({err: err.message});
-        }
-        process.send({xml: xml});
-    });
+
+    try {
+        var renderer = new InlineRenderer(input.ctx.params, input.ctx.options);
+
+        _.extend(renderer, input.ctx);
+
+        renderer.toXML(function (err, xml) {
+            if (err) {
+                return process.send({err: err.message});
+            }
+            process.send({xml: xml});
+        });
+    } catch (err) {
+        return process.send({err: err.message});
+    }
 });
